fix: 프로필 등록 안된 페이지
