+++ src/pages/profile/[id]/index.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const ProfileDetail = () => {
+  return <div>index</div>;
+};
+
+export default ProfileDetail;
